requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.1
httpx>=0.27.0
pandas>=2.2.0
numpy>=1.26.0

# Quantitative & Backtesting
yfinance>=0.2.37
vectorbt>=0.26.0
pandas-ta==0.4.71b0
statsmodels>=0.14.1
PyPortfolioOpt>=1.5.5

# LangChain & AI
langgraph>=0.0.30
langgraph-checkpoint-postgres>=1.0.0
langchain>=0.1.13
langchain-google-genai>=1.0.1
langchain-core>=0.1.33

# Microservices & Redis Cache
celery>=5.3.6
redis>=5.0.3
psycopg2-binary>=2.9.9

# API Clients & Utilities
tenacity>=8.2.3
beautifulsoup4>=4.12.3
lxml>=5.1.0
alpaca-trade-api>=3.1.1
